+++ web_customizer/static/src/customization_banner.js
@@ -1,7 +1,7 @@
+import {ROOT_OPTION_VIEW_TYPES, useCustomizationService} from "./customization_service";
 import {Component} from "@odoo/owl";
 import {_t} from "@web/core/l10n/translation";
 import {registry} from "@web/core/registry";
-import {useCustomizationService} from "./customization_service";
 
 export class CustomizationBanner extends Component {
     static template = "web_customizer.Banner";
@@ -21,6 +21,33 @@ export class CustomizationBanner extends Component {
 
     get exitLabel() {
         return _t("Exit");
+    }
+
+    get viewOptionsLabel() {
+        return _t("View options");
+    }
+
+    get rootTarget() {
+        const view = this.customization.state.view;
+        if (!view || !ROOT_OPTION_VIEW_TYPES.includes(view.viewType)) {
+            return null;
+        }
+        return view;
+    }
+
+    openViewOptions() {
+        const view = this.rootTarget;
+        if (!view) {
+            return;
+        }
+        this.customization.openFieldDialog({
+            fieldName: "",
+            fieldLabel: view.viewType,
+            model: view.model,
+            viewId: view.viewId,
+            viewType: view.viewType,
+            anchorKind: "view",
+        });
     }
 
     exit() {